+++ packages/data-explorer/src/components/Utils/ShoppingButton.vue
@@ -1,5 +1,6 @@
 <template>
   <button
+    @click="toggleShoppingItems(id)"
     type="button"
     :class="{'btn-primary': isSelected, 'btn-outline-secondary': !isSelected}"
     class="m-2 btn-sm float-right text-nowrap">
@@ -14,16 +15,24 @@ import Vue from 'vue'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faShoppingCart, faMinus, faPlus } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+import { mapMutations } from 'vuex'
 
 library.add(faShoppingCart, faMinus, faPlus)
 
 export default Vue.extend({
   name: 'ShoppingButton',
   props: {
+    id: {
+      type: String,
+      required: true
+    },
     isSelected: {
       type: Boolean,
       required: true
     }
+  },
+  methods: {
+    ...mapMutations(['toggleShoppingItems'])
   },
   components: { FontAwesomeIcon }
 })
