+++ packages/data-explorer/tests/unit/components/DataView/ExplorerCard.spec.ts
@@ -1,0 +1,9 @@
+import { shallowMount } from '@vue/test-utils'
+import EntityCard from '@/components/DataView/ExplorerCard.vue'
+
+describe('EntityCard.vue', () => {
+  it('exists', () => {
+    const wrapper = shallowMount(EntityCard, { propsData: { id: 'none' } })
+    expect(wrapper.exists()).toBeTruthy()
+  })
+})
