+++ packages/data-explorer/tests/unit/components/DataView/TableRow.spec.ts
@@ -1,5 +1,5 @@
 import { shallowMount } from '@vue/test-utils'
-import EntityTableRow from '@/components/DataView/EntityTableRow.vue'
+import EntityTableRow from '@/components/DataView/TableRow.vue'
 
 describe('EntityTableRow.vue', () => {
   it('exists', () => {
